Algorithm header added; loop variable type changed

+++ src/functions.cpp
@@ -1,9 +1,10 @@
 #include "../inc/functions.hpp"    // TODO: add proper compilation flag and change to #include "functions.hpp"
 #include <iostream>
 #include <limits>
+#include <algorithm>
 
 void print(std::vector<int> const & numbers) {
-    for (int i = 0; i < numbers.size(); ++i) {
+    for (size_t i = 0; i < numbers.size(); ++i) {
         std::cout << i << ": " << numbers[i] << '\n';
     }
     std::cout << '\n';
